Moved Search Code
